+++ frontend/app/src/main/java/de/amosproj3/ziofa/ui/visualization/composables/SelectMetricPrompt.kt
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2025 Luca Bretting <luca.bretting@fau.de>
+//
+// SPDX-License-Identifier: MIT
+
+package de.amosproj3.ziofa.ui.visualization.composables
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SelectMetricPrompt() {
+    CenteredInfoText(text = "Please make a valid selection.")
+}
